third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-in-older-shadow.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<div id="sandbox"></div>
<script>
var borderColor;

function borderColorOf(node)
{
    return document.defaultView.getComputedStyle(node, null).getPropertyValue('border-color');
}

function borderColorShouldBe(selector, color)
{
    var text = 'borderColorOf(getNodeInTreeOfTrees("' + selector + '"))';
    shouldBeEqualToString(text, color);
}

function cleanUp()
{
    document.getElementById('sandbox').innerHTML = '';
}

description('crbug.com/353406: Test for rules with ::content declared in older shadow trees');

var sandbox = document.getElementById('sandbox');
sandbox.appendChild(
    createDOM('div', {'id': 'host'},
        createShadowRoot(
            createDOM('style', {},
                document.createTextNode('content::content div { border: 1px solid green; }'),
            createDOM('div', {},
                document.createTextNode('inert')))),
        createShadowRoot(
            createDOM('content')),
        createDOM('div', {'id': 'target'},
            document.createTextNode('no border'))));

borderColorShouldBe('target', 'rgb(0, 128, 0)');

cleanUp();
</script>
</body>
</html>
