third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-in-older-shadow-expected.txt

crbug.com/353406: Test for rules with ::content declared in older shadow trees

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS borderColorOf(getNodeInTreeOfTrees("target")) is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
